move peer state management into a dedicated file
