+++ lib/storage/src/content_manager/toc/create_collection.rs
@@ -5,7 +5,7 @@ use collection::config::{self, CollectionConfigInternal, CollectionParams, Shard
 use collection::operations::config_diff::DiffConfig as _;
 use collection::operations::types::{CollectionResult, VectorsConfig};
 use collection::shards::collection_shard_distribution::CollectionShardDistribution;
-use collection::shards::replica_set::ReplicaState;
+use collection::shards::replica_set::replica_set_state::ReplicaState;
 use collection::shards::shard::{PeerId, ShardId};
 
 use super::TableOfContent;
