+++ lib/storage/src/content_manager/toc/snapshots.rs
@@ -2,7 +2,7 @@ use std::path::{Path, PathBuf};
 
 use collection::common::snapshots_manager::SnapshotStorageManager;
 use collection::operations::snapshot_ops::SnapshotDescription;
-use collection::shards::replica_set::ReplicaState;
+use collection::shards::replica_set::replica_set_state::ReplicaState;
 use collection::shards::shard::{PeerId, ShardId};
 use collection::shards::transfer::{ShardTransfer, ShardTransferMethod};
 use fs_err::tokio as tokio_fs;
