+++ src/common/metrics.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use api::rest::models::HardwareUsage;
-use collection::shards::replica_set::ReplicaState;
+use collection::shards::replica_set::replica_set_state::ReplicaState;
 use itertools::Itertools;
 use prometheus::TextEncoder;
 use prometheus::proto::{Counter, Gauge, LabelPair, Metric, MetricFamily, MetricType};
